Due to the overwhelming lack of responses I'll just commit this. Changes:
- Even more warning fixes (gcc 4 introduced a whole lot of new and exciting
ones). Some of those change how the code works.

- Fix a crash when saving without available sound device.

- Fix a division by zero crash.

- Move some stuff in the Makefiles around, and change CPP (which is used for
the preprocessor) to CC.

- Put the revision number into the binary. This breaks the Windows build. To
fix it, add a file src/revision.h containing '#define SVN_REVISION "$WCREV$"'
and call SubWCRev.exe on that file when building. For details see
http://subversion.tigris.org/faq.html#version-value-in-source

Yeah, blame me.



git-svn-id: svn+ssh://svn.gna.org/svn/warzone/branches/crossplatform@184 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/hci.c
@@ -6414,7 +6414,7 @@ static BOOL _intAddStats(BASE_STATS **ppsStatsList, UDWORD numStats,
 								sLabInit.y = STAT_BUTHEIGHT - 19;
 								sLabInit.width = 12;
 								sLabInit.height = 15;
-								sLabInit.pUserData = (void*)ii;
+								sLabInit.pUserData = (void*)(int)ii;
 								sLabInit.pTip = getPlayerName(ii);
 								sLabInit.pDisplay = intDisplayAllyIcon;
 								widgAddLabel(psWScreen, &sLabInit);
